requirements.txt
numpy
scipy
sounddevice
pyaudio
pvporcupine
TTS
whisper
resemblyzer
faiss-cpu
tinydb
pyAudioAnalysis
eyed3
python-dotenv
torch
torchaudio
opencv-python
matplotlib
scikit-image
scikit-learn
transformers
torch
accelerate
sentencepiece
